# docs/bin/openapi.json
{
    "components": {
...
        "schemas": {
...
            "ErrorShape": {
                "properties": {
                    "reason": {
                        "type": "string"
                    },
                    "code": {
                        "type": "number",
                        "format": "double"
                    }
                },
                "required": [
                    "reason",
                    "code"
                ],
                "type": "object"
            },
            "BlockSubset": {
                "properties": {
                    "slot": {
                        "type": "number",
                        "format": "double",
                        "example": 4924800
                    },
                    "epoch": {
                        "type": "number",
                        "format": "double",
                        "example": 209
                    },
                    "height": {
                        "type": "number",
                        "format": "double",
                        "example": 4512067
                    },
                    "hash": {
                        "type": "string",
                        "description": "[0-9a-fA-F]{64}",
                        "example": "cf8c63a909d91776e27f7d05457e823a9dba606a7ab499ac435e7904ee70d7c8"
                    },
                    "era": {
                        "type": "number",
                        "format": "double",
                        "example": 1
                    }
                },
                "required": [
                    "slot",
                    "epoch",
                    "height",
                    "hash",
                    "era"
                ],
                "type": "object"
            },
...
            "BlockInfo": {
                "allOf": [
                    {
                        "$ref": "#/components/schemas/BlockSubset"
                    },
                    {
                        "properties": {
                            "isValid": {
                                "type": "boolean"
                            },
                            "indexInBlock": {
                                "type": "number",
                                "format": "double",
                                "description": "index of tx in block"
                            }
                        },
                        "required": [
                            "isValid",
                            "indexInBlock"
                        ],
                        "type": "object"
                    }
                ]
            },
...
            "UtxoPointer": {
                "properties": {
                    "index": {
                        "type": "number",
                        "format": "double"
                    },
                    "txHash": {
                        "type": "string",
                        "example": "011b86557367525891331b4bb985545120efc335b606d6a1c0d5a35fb330f421",
                        "pattern": "[0-9a-fA-F]{64}"
                    }
                },
                "required": [
                    "index",
                    "txHash"
                ],
                "type": "object"
            },
            "UtxoAndBlockInfo": {
                "properties": {
                    "utxo": {
                        "allOf": [
                            {
                                "$ref": "#/components/schemas/UtxoPointer"
                            },
                            {
                                "properties": {
                                    "payload": {
                                        "type": "string",
                                        "example": "825839019cb581f4337a6142e477af6e00fe41b1fc4a5944a575681b8499a3c0bd07ce733b5911eb657e7aff5d35f8b0682fe0380f7621af2bbcb2f71b0000000586321393",
                                        "pattern": "[0-9a-fA-F]*"
                                    }
                                },
                                "required": [
                                    "payload"
                                ],
                                "type": "object"
                            }
                        ]
                    },
                    "block": {
                        "$ref": "#/components/schemas/BlockInfo"
                    }
                },
                "required": [
                    "utxo",
                    "block"
                ],
                "type": "object"
            },
            "TransactionOutputResponse": {
                "properties": {
                    "utxos": {
                        "items": {
                            "$ref": "#/components/schemas/UtxoAndBlockInfo"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "utxos"
                ],
                "type": "object"
            },
            "TransactionOutputRequest": {
                "properties": {
                    "utxoPointers": {
                        "items": {
                            "$ref": "#/components/schemas/UtxoPointer"
                        },
                        "type": "array"
                    }
                },
                "required": [
                    "utxoPointers"
                ],
                "type": "object"
            }
        },
        "securitySchemes": {}
    },
    "info": {
        "title": "carp",
        "version": "1.0.0",
        "description": "API for the Postgres database generated by Carp",
        "license": {
            "name": "MIT"
        },
        "contact": {
            "name": "https://github.com/dcSpark"
        }
    },
    "openapi": "3.0.0",
    "paths": {
...
        "/transaction/output": {
            "post": {
                "operationId": "TransactionOutput",
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/TransactionOutputResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorShape"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ErrorShape"
                                }
                            }
                        }
                    }
                },
                "description": "Get the outputs for given `<tx hash, output index>` pairs\n\nNote: this endpoint only returns txs that are in a block. Use another tool to see mempool for txs not in a block",
                "security": [],
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/TransactionOutputRequest"
                            }
                        }
                    }
                }
            }
        }
    },
    "servers": [
        {
            "url": "/"
        }
    ]
}